+++ tests/test_connection.py
@@ -40,6 +40,13 @@ async def test_type_success(pair_factory, expect_codes_in_exception):
 
 
 @pytest.mark.asyncio
+async def test_custom_passive_commands(pair_factory):
+    async with pair_factory(host="127.0.0.1") as pair:
+        pair.client._passive_commands = None
+        await pair.client.get_passive_connection("A", commands=["pasv", "epsv"])
+
+
+@pytest.mark.asyncio
 async def test_extra_pasv_connection(pair_factory):
     async with pair_factory() as pair:
         r, w = await pair.client.get_passive_connection()